Add files via upload
This is the first version of code using the fixed template code supplied for the NOMAD data. The initial target here is to create smaller plots of certain ranges of solar longitude corresponding to the equinoxes and solstices as well as aphelion and perihelion between 163 Ls in MY34 and the end of MY35. 

This initial code has created a plot for the Autumnal Equinox in MY34, in a range of 20 solar longitudes either side of the equinox. I have also developed a line of code to automate the labels of the x-axis so this can be reused for each of the other 9 timeframes I will be creating plots for.