Add files via upload
Work from the last few weeks. In this code I retrieve data from the MCD for a specific altitude and latitude, specifically for temperature. I create a plot with it and then create code to make a GIF for all the solar longitudes, 0 to 360. I forogt to checkpoint each change, but it can mostly be observed through the cells. I have further jupyter notebooks to upload detailing more of these changes.