Add files via upload
This is the first version of the final code (for gathering data from the MCD to compare with NOMAD). I am currently working on iterations for latitude, altitude, and different variables to compare. for each of these, the links will be stored in an array in a dictionary, as well as gifs being created for each of these over 360 solar longitudes.