Add files via upload
This is the cleaner code (without unnecessary commented previous code etc). 
The main problem I have been having is with reversing the code for zz. I have now decided to use numerical method to calculate it, from the Earth to Mars date code. 

The reason I have been doing this is because the Mars Year matters on the MCD, as you get different model data for each year. If not using the exact same times for the model and the NOMAD data, the evidence and conclusions drawn would be ineffective.